Don't install sass globally
There is no need any more to install a specific sass version because
the user debian distribution will allow installing sass in the required
version without having to use gcc to compile stuff etc. So we can have
a generic image with the tools but without the globally installed sass
package.